Extract SseSessionRegistry into its own translation unit  (#215)
Pull the SSE session-id-to-connection map out of the HTTP+SSE filter
chain factory's .cc and into include/mcp/filter/sse_session_registry.h
+ src/filter/sse_session_registry.cc. No behavior change; the factory
includes the header and uses the class exactly as before.

The extraction is driven by testability: as a file-local class the
registry could only be exercised indirectly through a full filter
chain, which meant there was no narrow unit test for the tiny piece
of shared state responsible for routing every POST /callback response.
Giving it a real header also makes the class reusable if another MCP
transport ever needs the same session-id-to-stream mapping pattern
(the registry itself is transport-agnostic).

Also add two small introspection accessors (sessionCount, hasSession)
that tests use to assert state after a sequence of
register/removeSession calls, without reaching into private members.
